feat: move last files to dedicated dir
