+++ libs/coin-polkadot/src/bridge/estimateMaxSpendable.ts
@@ -2,10 +2,10 @@ import { BigNumber } from "bignumber.js";
 import type { AccountLike, Account } from "@ledgerhq/types-live";
 import { getMainAccount } from "@ledgerhq/coin-framework/account/index";
 import type { PolkadotAccount, Transaction } from "../types";
-import { calculateAmount } from "../logic";
+import { calculateAmount } from "./logic";
 import getEstimatedFees from "./getFeesForTransaction";
 import createTransaction from "./createTransaction";
-import { loadPolkadotCrypto } from "../polkadot-crypto";
+import { loadPolkadotCrypto } from "./polkadot-crypto";
 import { PolkadotAPI } from "../api";
 
 /**
