+++ libs/coin-polkadot/src/bridge/index.ts
@@ -16,16 +16,16 @@ import type {
   SignedOperation,
 } from "@ledgerhq/types-live";
 import { PolkadotAPI } from "../api";
-import resolver from "../hw-getAddress";
+import resolver from "./hw-getAddress";
 import createTransaction from "./createTransaction";
 import estimateMaxSpendable from "./estimateMaxSpendable";
 import getTransactionStatus from "./getTransactionStatus";
 import prepareTransaction from "./prepareTransaction";
 import buildSignOperation from "./signOperation";
 import { makeGetAccountShape } from "./synchronisation";
-import { loadPolkadotCrypto } from "../polkadot-crypto";
-import { assignFromAccountRaw, assignToAccountRaw } from "../serialization";
-import { getPreloadStrategy, hydrate, preload } from "../preload";
+import { loadPolkadotCrypto } from "./polkadot-crypto";
+import { assignFromAccountRaw, assignToAccountRaw } from "./serialization";
+import { getPreloadStrategy, hydrate, preload } from "./preload";
 import type { PolkadotAddress, PolkadotSignature, Transaction } from "../types";
 import { PolkadotSigner } from "../types";
