+++ libs/coin-polkadot/src/bridge/signOperation.ts
@@ -19,7 +19,7 @@ import type {
 import { encodeOperationId } from "@ledgerhq/coin-framework/operation";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import { buildTransaction } from "./buildTransaction";
-import { calculateAmount, getNonce, isFirstBond } from "../logic";
+import { calculateAmount, getNonce, isFirstBond } from "./logic";
 import { PolkadotAPI } from "../api";
 import { PolkadotAddress, PolkadotSignature, PolkadotSigner } from "../types";
