+++ libs/coin-polkadot/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./test";
 export * from "./types";
 
 export { buildCurrencyBridge, buildAccountBridge, createBridges } from "./bridge/index";
