+++ libs/coin-polkadot/src/test/index.ts
@@ -1,0 +1,4 @@
+export * from "./bridge.integration.test";
+export * from "./cli";
+export * from "./specs";
+export * from "./speculos-deviceActions";
